requirements.txt
chromium-compact-language-detector==0.31415
Django==1.8.5
django-ajax-selects==1.4.1
django-betterforms==1.1.2
django-filter==0.12.0
django-mptt==0.7.4
django-post-office==2.0.2
django-timezone-field==1.2
djangorestframework==3.3.2
gdata==2.0.18
httplib2==0.9.2
jsonfield==1.0.3
Markdown==2.6.5
oauth2client==1.5.1
psycopg2==2.6.1
pyasn1==0.1.9
pyasn1-modules==0.0.8
python-dateutil==2.4.2
python-social-auth==0.2.19
pytz==2015.6
rsa==3.2
six==1.9.0
